+++ db/migrate/20161026132144_add_short_name_to_designs.rb
@@ -1,0 +1,5 @@
+class AddShortNameToDesigns < ActiveRecord::Migration[5.0]
+  def change
+    add_column :designs, :short_name, :string
+  end
+end